[UPDATE][REFACTOR] Made gf_coefficients more general, added Sage script
The routines in gf_coefficients are now -- as much as possible --
independent from the actual choice of the Galois-field. We added a
Sage script and templates to generate the gf_coefficients.h and
gf_coefficients.c from almost any applicable finite field. This gives
an automatized way to change the GF inside our project to be as
arbitrary as possible.

NOTE: the tests are not yet made such that they are true for all
Galois fields that can be defined. For now, the tests only apply to
GF(125).